+++ src/design/plone/contenttypes/browser/manage_content/check_servizi.py
@@ -72,6 +72,15 @@ class CheckServizi(BrowserView):
             "ufficio_responsabile": getattr(servizio, "ufficio_responsabile", None),
         }
 
+    def plone2volto(self, url):
+        portal_url = api.portal.get().absolute_url()
+        frontend_domain = api.portal.get_registry_record(
+            "volto.frontend_domain", default=""
+        )
+        if frontend_domain and url.startswith(portal_url):
+            return url.replace(portal_url, frontend_domain, 1)
+        return url
+
     def get_servizi(self):
         if self.is_anonymous():
             return []
@@ -97,14 +106,14 @@ class CheckServizi(BrowserView):
             parent = servizio.aq_inner.aq_parent
             if parent.title not in results:
                 results[parent.title] = {
-                    "url": parent.absolute_url().replace("/api/", "/"),
+                    "url": self.plone2volto(parent.absolute_url()),
                     "children": [],
                 }
 
             results[parent.title]["children"].append(
                 {
                     "title": servizio.title,
-                    "url": servizio.absolute_url().replace("/api/", "/"),
+                    "url": self.plone2volto(servizio.absolute_url()),
                     "data": {
                         "title": information_dict.get("title") and FLAG or "",
                         "description": information_dict.get("description")